+++ src/Order/Contracts/FulfillmentStatus.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the FulfillmentStatus interface.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-02-20
+ *
+ */
+
+namespace Vanilo\Order\Contracts;
+
+interface FulfillmentStatus
+{
+}