+++ src/Order/Models/FulfillmentStatusProxy.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the FulfillmentStatusProxy class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-02-20
+ *
+ */
+
+namespace Vanilo\Order\Models;
+
+use Konekt\Concord\Proxies\EnumProxy;
+
+class FulfillmentStatusProxy extends EnumProxy
+{
+}