+++ src/Order/Providers/ModuleServiceProvider.php
@@ -20,6 +20,7 @@ use Vanilo\Order\Contracts\OrderFactory as OrderFactoryContract;
 use Vanilo\Order\Contracts\OrderNumberGenerator;
 use Vanilo\Order\Factories\OrderFactory;
 use Vanilo\Order\Models\Billpayer;
+use Vanilo\Order\Models\FulfillmentStatus;
 use Vanilo\Order\Models\Order;
 use Vanilo\Order\Models\OrderItem;
 use Vanilo\Order\Models\OrderStatus;
@@ -29,11 +30,12 @@ class ModuleServiceProvider extends BaseModuleServiceProvider
     protected $models = [
         Billpayer::class,
         Order::class,
-        OrderItem::class
+        OrderItem::class,
     ];
 
     protected $enums = [
-        OrderStatus::class
+        OrderStatus::class,
+        FulfillmentStatus::class,
     ];
 
     public function boot()