+++ src/Order/resources/database/migrations/2023_02_20_145123_add_fulfillment_status_to_orders.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+return new class () extends Migration {
+    public function up()
+    {
+        Schema::table('orders', function (Blueprint $table) {
+            $table->string('fulfillment_status')->default('unfulfilled')->after('status');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('orders', function (Blueprint $table) {
+            $table->dropColumn('fulfillment_status');
+        });
+    }
+};